Add new implementations and updates across various files
- Implemented a new Python script for Advent of Code 2025 Day 1 challenge with functions for part 1 and part 2.
- Modified C++ ArrayList implementation to include math header and corrected function return.
- Updated Python PrefixSum script to include collections import.
- Removed unused TypeScript and Dart files from the Random directory.
- Added input file for tree structure representation.
- Created new C++ files for CSES problems including Array Manipulation and Static Range Sum Query with type adjustments.
- Added Python binary search implementation for CSES.
- Introduced new C program demonstrating pointer usage with strings.
- Added binary file for CSES solutions.
- Implemented two different heap structures in Python with methods for insertion, polling, and heapifying.

+++ cses/StaticRangeSumQuery.cpp
@@ -3,20 +3,20 @@
 #define Log(x) for(auto e : x){ std::cout << e << " ";} std::cout << "\n";
 
 int main(){
-    long long n, q;
+    long n, q;
     std::cin >> n;
     std::cin >> q;
-    std::vector<long long> v;
-    std::vector<std::pair<long long, long long >> pairs;
+    std::vector<long> v;
+    std::vector<std::pair<long, long>> pairs;
     std::vector<long> prefix(n + 1);
     for(long i = 0; i < n; i++ ){
-        long long val;
+        long val;
         std::cin >> val;
         v.push_back(val);
     }
 
     for(long i = 0; i < q; i++){
-        long long first, second;
+        long first, second;
         std::cin >> first;
         std::cin >> second;
         pairs.push_back(std::pair<int,int>(first, second));
